Ajusted visuales primer despliegue

+++ src/views/RHomeView.js
@@ -1,4 +1,3 @@
-import api from "../api/axiosConfig";
 import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
@@ -23,7 +22,6 @@ function RHomeView() {
   const navigate = useNavigate();
 
   const [showPopup, setShowPopup] = useState(false);
-  const [isNewStory, setIsNewStory] = useState(false);
 
   useEffect(() => {
     dispatch(setNarrative(null));
@@ -43,19 +41,13 @@ function RHomeView() {
   };
 
   const newStory = () => {
-    const id_usuario = localStorage.getItem('id_usuario');
-
-    api.get(`/historias/${id_usuario}`)
-      .then(res => setIsNewStory(res.data.length === 0 ? true : false))
-      .catch(err => console.error('Error al cargar las historias del usuario:', err));
-
-    if(isNewStory)
-      navigate('/narratives', { state: { isNewStory } })
-    else
-      navigate('/narratives')
+    navigate('/narratives')
   };
 
   const handleLogout = () => {
+    const confirmar = window.confirm("¿Seguro de que quieres cerrar sesión?");
+    if (!confirmar) return;
+
     persistor.purge();
     localStorage.clear();
     sessionStorage.clear();
